refactor: 입력 시 예외 처리 로직 별도의 클래스 분리
